requirements.txt
praw==7.7.0
requests==2.31.0
urllib3==2.0.5
pymongo==4.5.0
python-dotenv==1.0.0
openai==0.27.8
